+++ lib/cri/commands/basic_root.rb
@@ -1,7 +1,7 @@
 # encoding: utf-8
 
-option :h, :help, 'show help for this command' do |value|
-  puts self.help
+flag :h, :help, 'show help for this command' do |value, cmd|
+  puts cmd.help
   exit 0
 end