+++ NEWS.md
@@ -1,6 +1,12 @@
 Cri News
 ========
 
+2.3.0
+-----
+
+* Added colors
+* Added support for marking commands as hidden
+
 2.2.1
 -----